separated spec files for spacer
