+++ cypress/integration/v2/blocks/spacer/other-tests.spec.js
@@ -1,0 +1,17 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import {
+	blockExist,
+	blockError,
+} from '.'
+
+describe( 'Spacer Block ( Other Tests )', registerTests( [
+	blockExist,
+	blockError,
+] ) )
